refactor(sema): split natspec helpers
