+++ crates/sema/src/ast_lowering/lower.rs
@@ -3,6 +3,7 @@ use solar_ast as ast;
 use solar_data_structures::smallvec::SmallVec;
 
 mod natspec;
+mod signature;
 
 impl<'gcx> super::LoweringContext<'gcx> {
     #[instrument(level = "debug", skip_all)]
